fix(time_series): Update fp_forecast_mape signature from int to size_t
Change fp_forecast_mape parameter from 'int n' to 'size_t n' for
consistency with fp_forecast_mae and fp_forecast_rmse.

Updated in both:
- Implementation: src/algorithms/fp_time_series.c
- Forward declaration: examples/algorithms/demo_time_series.c

+++ src/algorithms/fp_time_series.c
@@ -581,10 +581,10 @@ void fp_generate_random_walk(
 // ============================================================================
 
 // Mean Absolute Percentage Error
-double fp_forecast_mape(const double* actual, const double* predicted, int n) {
+double fp_forecast_mape(const double* actual, const double* predicted, size_t n) {
     double sum_ape = 0.0;
-    int count = 0;
-    for (int i = 0; i < n; i++) {
+    size_t count = 0;
+    for (size_t i = 0; i < n; i++) {
         if (fabs(actual[i]) > 1e-10) {  // Avoid division by zero
             sum_ape += fabs((actual[i] - predicted[i]) / actual[i]);
             count++;
